trampoline: Correctly set boot CPU on little-endian
Currently, a little-endian trampoline will store the boot cpu value
into the device-tree header with a simple stw. However, the fdt is
defined as big-endian, so we'll be setting an endian-reversed value
here.

This change uses a stwbrx on little-endian, and a stwx for big-endian.
Because we only have an x-form instruction for a byte-reversed store, we
need an extra instruction to add our offset.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>