Don't always update initrd location in the device-tree
Currently kexec-lite always updates the location of the initrd
(linux,initrd-start and linux,initrd-end) in the device-tree it intends
to pass to the next kernel.

This behaviour means that if no --initrd argument is passed,
linux,initrd-start and linux,initrd-end get zeroed. Zeroing those fields
prevents the next kernel from reusing an initrd in ram that the first
kernel was using.