Use a fixed reserve rather than up to linux,kernel-end
Currently we reserve the memory from 0...linux,kernel-end in the kexec
memory map which prevents the current kernel's text from being
overwritten when we assemble the kexec segment buffers at kexec time.

This reservation is only really needed when loading a crashkernel since
it wants to preserve the existing kernel text to create a crashdump. For
a generic vmlinux the reservation is mostly pointless since most
kernel's will copy themselves down to in early boot anyway.

When a zImage is loaded we do need to keep the start of memory reserved
because the zImage will extract the vmlinux to zero on ePAPR platforms.
By luck reserving up until linux,kernel-end will usually reserve enough
space for the new kernel's text. This does however break down when the
new kernel is signifigantly larger than the previous (e.g. it contains a
large builtin initramfs).

This patch replaces the 0...linux,kernel-end reservation with a fixed
256MB reservation, which ought to be enough for anybody.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>